Redesigned Layout, Renamed Classes
Redesigned layout to roughly match new design.

Refactored and renamed classes to be more descriptive.
