+++ src/main/resources/org/example/Orders.fxml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+
+<AnchorPane fx:id="anchorPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="740.0" prefWidth="960.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.OrdersController">
+   <children>
+      <StackPane layoutX="157.0" layoutY="160.0" prefHeight="740.0" prefWidth="637.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="165.0" AnchorPane.rightAnchor="165.0" AnchorPane.topAnchor="0.0">
+         <children>
+            <VBox alignment="CENTER" fillWidth="false" prefHeight="200.0" prefWidth="1010.0" spacing="10.0">
+               <children>
+                  <Label text="Enter a new order below:" />
+                  <TextField fx:id="orderNumTextField" prefHeight="25.0" prefWidth="192.0" promptText="Enter an order number" />
+                  <TextField promptText="Test field (ignore)" />
+                  <HBox alignment="CENTER" prefHeight="44.0" prefWidth="207.0" spacing="25.0">
+                     <children>
+                        <Button fx:id="cancelButton" mnemonicParsing="false" onAction="#switchToPrimary" text="Cancel" />
+                        <Button fx:id="saveOrderButtoin" mnemonicParsing="false" onAction="#saveOrder" text="Save" />
+                     </children>
+                     <padding>
+                        <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
+                     </padding>
+                  </HBox>
+                  <Label fx:id="testLabel" text="Label" />
+               </children>
+               <padding>
+                  <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+               </padding>
+            </VBox>
+         </children>
+      </StackPane>
+      <Button fx:id="backButton" layoutX="14.0" layoutY="14.0" mnemonicParsing="false" onAction="#switchToPrimary" prefHeight="33.0" prefWidth="145.0" text="Back to Dashboard" />
+   </children>
+</AnchorPane>
