--- src/main/resources/org/example/primary.fxml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.TilePane?>
-<?import javafx.scene.layout.VBox?>
-
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="740.0" prefWidth="960.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.PrimaryController">
-   <left>
-      <VBox alignment="TOP_CENTER" prefHeight="695.0" prefWidth="160.0" spacing="20.0" style="-fx-background-color: gray;" BorderPane.alignment="CENTER">
-         <children>
-            <Button fx:id="primaryButton" minWidth="-Infinity" mnemonicParsing="false" onAction="#switchToSecondary" prefHeight="25.0" prefWidth="134.0" text="Enter a New Order">
-               <VBox.margin>
-                  <Insets />
-               </VBox.margin>
-            </Button>
-            <Button fx:id="addOrderButton" mnemonicParsing="false" onAction="#addOrderWindow" text="Add Order (window-test)" />
-            <Button mnemonicParsing="false" onAction="#printNameFromEntry" text="Print Num (testing)" />
-            <TextField fx:id="testTextField" maxWidth="-Infinity" prefHeight="25.0" prefWidth="162.0" promptText="Enter a number (testing)" />
-         </children>
-         <padding>
-            <Insets bottom="20.0" left="5.0" right="5.0" top="20.0" />
-         </padding>
-      </VBox>
-   </left>
-   <top>
-      <TilePane prefHeight="45.0" prefWidth="960.0" style="-fx-background-color: #42aaf5;" BorderPane.alignment="CENTER" />
-   </top>
-   <center>
-      <VBox alignment="CENTER" spacing="8.0" BorderPane.alignment="CENTER">
-         <children>
-            <Label text="Newest Order:" />
-            <HBox fillHeight="false" maxWidth="-Infinity" prefHeight="21.0" prefWidth="230.0" spacing="10.0">
-               <children>
-                  <Label fx:id="orderNumLabel" text="Last Name:          " />
-                  <Label fx:id="lastNameLabel" />
-               </children>
-               <padding>
-                  <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-               </padding>
-            </HBox>
-            <HBox fillHeight="false" layoutX="295.0" layoutY="357.0" maxWidth="-Infinity" prefHeight="21.0" prefWidth="230.0" spacing="10.0">
-               <children>
-                  <Label fx:id="orderNumLabel1" text="First Name:          " />
-                  <Label fx:id="firstNameLabel" />
-               </children>
-               <padding>
-                  <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-               </padding>
-            </HBox>
-            <HBox fillHeight="false" layoutX="295.0" layoutY="374.0" maxWidth="-Infinity" prefHeight="21.0" prefWidth="230.0" spacing="10.0">
-               <children>
-                  <Label fx:id="orderNumLabel11" text="Invoice Number:  " />
-                  <Label fx:id="invoiceLabel" />
-               </children>
-               <padding>
-                  <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-               </padding>
-            </HBox>
-         </children>
-         <padding>
-            <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-         </padding>
-      </VBox>
-   </center>
-</BorderPane>
